feat: Add abstract class for LibraryItem
This commit adds an abstract class called LibraryItem to the LibraryManagementSystem.cpp file. The LibraryItem class provides a common interface for different types of library items and includes member variables for title and id. It also includes a pure virtual function called display() which will be implemented by derived classes. This abstract class will serve as the base for future library item implementations.