feat: Add loan functionality for books
The commit adds loan functionality for books in the LibraryManagementSystem.cpp file. It includes the creation of book objects, patron objects, and a LoanManager object. The loanItem() function is used to loan books to patrons, and the returnItem() function is used to return books. The displayLoanedItems() function is used to display the currently loaned books. This commit implements the LoanManager class and allows for the management of loaned items in the library management system.